Refactor: move dropdown logic to use select goal menu hook
